# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ocolis",
    "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. In quis aliquam neque. Maecenas non pretium quam. Etiam sapien metus, bibendum eget neque nec, luctus maximus lacus. Vestibulum bibendum ipsum at ante mollis, id tristique massa eleifend. Etiam eu ipsum eros. Fusce ac fringilla ligula. Aenean fermentum augue nisl, eu congue tellus fermentum ut. Sed sodales, est vel sagittis faucibus, lorem velit rutrum tellus, vitae sagittis ex erat nec odio. Nam consequat libero in ante feugiat, nec ornare nunc hendrerit.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:7070"
    }
  ],
...
  "paths": {
    "/user": {
      "get": {
        "tags": [
          "utilisateurs"
        ],
        "summary": "Récupère tout les utilisateurs stockés en base de données",
        "operationId": "getAllUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de la requête",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {},
                  "example": {
                    "id": 3,
                    "email": "gordon.hill@gmail.com",
                    "status": "user"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Le client n'est pas authentifié, le token n'est pas valide/expiré ou l'utilisateur n'existe plus"
          },
          "404": {
            "description": "Aucun utilisateur stocké en base de données"
          }
        }
      },
      "post": {
        "tags": [
          "utilisateurs"
        ],
        "summary": "Création d'un nouvel utilisateur",
        "operationId": "createUser",
        "requestBody": {
          "description": "Informations concernant le nouvel utilisateur. Le mail doit être valide et unique, le mot de passe doit contenir au moins 8 lettres dont une majuscule, une minuscule, un chiffre et un caractère spécial. Les deux champs sont requis",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "byte"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Utilisateur crée. Renvoie l'id du nouvel utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "example": 4
                }
              }
            }
          },
          "400": {
            "description": "Echec de la requête. Le mail indiqué est déjà utilisé ou invalide, un des deux champs obligatoires est manquant ou le mot de passe n'est pas suffisamment complexe"
          }
        }
      }
    },
    "/user/{userId}": {
      "get": {
        "tags": [
          "utilisateurs"
        ],
        "summary": "Récupère un utilisateur selon son id",
        "operationId": "getUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Id de l'utilisateur recherché",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de la requête",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": 2,
                    "email": "irene.Carroll@gmail.com",
                    "status": "admin"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Le client n'est pas authentifié, le token n'est pas valide/expiré ou l'utilisateur n'existe plus"
          },
          "404": {
            "description": "Utilisateur inexistant/Id incorrect"
          }
        }
      },
      "patch": {
        "tags": [
          "utilisateurs"
        ],
        "summary": "Modifie un utilisateur selon son id",
        "operationId": "updateUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Id de l'utilisateur à modifier",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Nouvelles informations pour l'utilisateur indiqué. Le mail doit être valide et unique, le mot de passe doit contenir au moins 8 lettres dont une majuscule, une minuscule, un chiffre et un caractère spécial",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "byte"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "status_id": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès de la requête. L'attribut \"modified\" permet d'indiquer si le changement a été effectif ou non",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "modified": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Echec de la requête. Le mail indiqué est déjà utilisé ou invalide, le mot de passe n'est pas suffisamment complexe ou le status_id n'est pas un entier"
          },
          "401": {
            "description": "Le client n'est pas authentifié, le token n'est pas valide/expiré ou l'utilisateur n'existe plus"
          },
          "403": {
            "description": "Opération réservée aux administrateurs ou utilisateur non autorisé à modifier son propre statut"
          },
          "404": {
            "description": "Utilisateur inexistant/Id incorrect"
          }
        }
      },
      "delete": {
        "tags": [
          "utilisateurs"
        ],
        "summary": "Supprime un utilisateur selon son id",
        "operationId": "deleteUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Id de l'utilisateur à supprimer",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de la requête. L'attribut \"deleted\" permet d'indiquer si le changement a été effectif ou non",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          },
          "401": {
            "description": "Le client n'est pas authentifié, le token n'est pas valide/expiré ou l'utilisateur n'existe plus"
          },
          "403": {
            "description": "Utilisateur non autorisé à supprimer un autre profil que le sien"
          },
          "404": {
            "description": "Utilisateur inexistant/Id incorrect"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "utilisateurs"
        ],
        "summary": "Connexion utilisateur",
        "operationId": "loginUser",
        "requestBody": {
          "description": "Informations de connexion pour l'utilisateur. Les deux champs sont requis",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "byte"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Succès de la requête. Renvoie l'id de l'utilisateur connecté ainsi qu'un token permettant la connexion via \"bearer token\"",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "userId": 3,
                    "token": "eyJherkoOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsInVzZXJTdGF0dXNfaWQiOjEsImlhdCI6MTYyODY5NDYzNH7.7GuaybGOWv0xldvnOGVBM-vOiSn5wN9HcRRJi4Xh67Q"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Echec de la requête. Au moins un des deux champs obligatoires est manquant"
          },
          "401": {
            "description": "Les informations de connexion ne sont pas valides ou l'utilisateur n'existe pas"
          }
        }
      }
    },
    "/user/logout": {
      "get": {
        "tags": [
          "utilisateurs"
        ],
        "summary": "Déconnexion utilisateur",
        "operationId": "logoutUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Succès de la requête, utilisateur déconnecté"
          },
          "401": {
            "description": "Le client n'est pas authentifié, le token n'est pas valide/expiré ou l'utilisateur n'existe plus"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}